Refactor: Extract XP-EHH analysis module
